+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 07-06-2016 - Tony Partridge
+* Updated hardcoded english for iCal Repeat Deleted and iCal Repeats Deleted.
+* Fixed undefined variable within iCalRepeat on redirect.
 * Fixed jevuser.php where $idsstring was undefined, it should have been $idstring
 * Replace JButton with JToolBarButton as JButton is not deprecated.
 * Declared editStrings variable in icalrepat/view.html.php to avoid a coding notice in code editor. Has not other effect but to be cleaner.